Do not initialize in tasks the Core by default, only when requested
