+++ classes/Task/Restore/RestoreFiles.php
@@ -124,9 +124,4 @@ class RestoreFiles extends AbstractTask
 
         return ExitCode::SUCCESS;
     }
-
-    public function init(): void
-    {
-        // Do nothing
-    }
 }
